daemon/pkg/registry: move searchRepositories to where it's used
