+++ daemon/pkg/registry/search_session.go
@@ -4,13 +4,9 @@ import (
 	// this is required for some certificates
 	"context"
 	_ "crypto/sha512"
-	"encoding/json"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/cookiejar"
-	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 
@@ -194,42 +190,4 @@ func authorizeClient(ctx context.Context, client *http.Client, authConfig *regis
 	client.Jar = jar
 
 	return nil
-}
-
-// defaultSearchLimit is the default value for maximum number of returned search results.
-const defaultSearchLimit = 25
-
-// searchRepositories performs a search against the remote repository
-func searchRepositories(ctx context.Context, client *http.Client, ep *v1Endpoint, term string, limit int) (*registry.SearchResults, error) {
-	if limit == 0 {
-		limit = defaultSearchLimit
-	}
-	if limit < 1 || limit > 100 {
-		return nil, invalidParamf("limit %d is outside the range of [1, 100]", limit)
-	}
-	u := ep.String() + "search?q=" + url.QueryEscape(term) + "&n=" + url.QueryEscape(strconv.Itoa(limit))
-	log.G(ctx).WithField("url", u).Debug("searchRepositories")
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u, http.NoBody)
-	if err != nil {
-		return nil, invalidParamWrapf(err, "error building request")
-	}
-	// Have the AuthTransport send authentication, when logged in.
-	req.Header.Set("X-Docker-Token", "true")
-	res, err := client.Do(req)
-	if err != nil {
-		return nil, systemErr{err}
-	}
-	defer res.Body.Close()
-	if res.StatusCode != http.StatusOK {
-		// TODO(thaJeztah): return upstream response body for errors (see https://github.com/moby/moby/issues/27286).
-		// TODO(thaJeztah): handle other status-codes to return correct error-type
-		return nil, errUnknown{fmt.Errorf("unexpected status code %d", res.StatusCode)}
-	}
-	result := &registry.SearchResults{}
-	err = json.NewDecoder(res.Body).Decode(result)
-	if err != nil {
-		return nil, systemErr{errors.Wrap(err, "error decoding registry search results")}
-	}
-	return result, nil
 }
